Moves FormFieldMapCoordinates
