+++ src/components/Field/FormFieldMapCoordinates/Map/ResetButton.tsx
