+++ src/components/Field/FormFieldMapCoordinates/MapCoordinates.stories.tsx
